replace screener with generic model name; change bigints to ints

+++ db/migrate/20251126182903_create_screeners.rb
@@ -1,8 +1,8 @@
 class CreateScreeners < ActiveRecord::Migration[8.0]
   def change
     create_table :screeners do |t|
-      t.bigint :language_preference_written, default: 0, null: false
-      t.bigint :language_preference_spoken, default: 0, null: false
+      t.integer :language_preference_written, default: 0, null: false
+      t.integer :language_preference_spoken, default: 0, null: false
 
       t.timestamps
     end
